주문 상태 변화 시 DB 필드값 변경 기능 구현 완료.
그러나, 상태변화시 버튼과 텍스트뷰의 색상이 계속 제대로 바뀌었다가
refresh되면서 다시 이전 색상으로 돌아가는 현상이 있다.
이부분을 해결해야함

삭제 명령 - 왼쪽으로 밀면 삭제 되도록 함.
단, 삭제 스와이프의 민감도가 높은 경향이 있다. ( 이부분 좀 수정)

오른쪽 스와이프 시 다른 명령 여부 확인

+++ Pear/app/src/main/java/com/greenhelix/pear/orderStatus/OrderStatusActivity.java
@@ -13,6 +13,7 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import com.firebase.ui.firestore.FirestoreRecyclerOptions;
 import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
 import com.greenhelix.pear.R;
@@ -69,20 +70,25 @@ public class OrderStatusActivity extends AppCompatActivity {
 
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-                adapter.deleteItem(viewHolder.getAdapterPosition()); //해당 위치 position이 viewholder에 있는 해당 adaterposition이다.
+                if(direction == ItemTouchHelper.LEFT){
+                    Log.d(LOG_TAG, "삭제명령");
+                    adapter.deleteItem(viewHolder.getAdapterPosition()); //해당 위치 position이 viewholder에 있는 해당 adaterposition이다.
+                }
             }
         }).attachToRecyclerView(cycleOrderStatusView); //마지막에는 순환뷰에 적용
+
 
         //보통은 해당 클릭리스너 만들면, this로 implement시켜서 사용하지만,
         //커스텀을 한상태로 클릭 리스너를 만들었기 때문에 아래와 같이 사용한다.
 //        adapter.setOnCardClickListener(new OrderStatusAdapter.OnOrderClickListener() {
 //            @Override
 //            public void onItemClick(DocumentSnapshot documentSnapshot, int position) {
-//                NowOrder nowOrder = documentSnapshot.toObject(NowOrder.class);
-//                String id = documentSnapshot.getId();
+//                NowOrder nowOrder = documentSnapshot.toObject(NowOrder.class);//nowOrder로 정보 변이가능
+//                String id = documentSnapshot.getId(); //문서의 아이디 값 가져오기
 //                Log.d(LOG_TAG, "activity로 잘 넘어왔다.\n 현재 선택한 문서의 아이디는 " + id);
 //            }
 //        });
+        //adapter.setOnCardClickListener여기까지는 그대로 쳐주고, 그다음 부분()안에 올 곳이 adapter에서 정의한 커스텀 onclick메서드이다.
     }
 
     @Override
